Update Arma3Profile.cs
cause double int don't make sense. I fixed it all by my self :p

+++ FASTER/Models/Arma3Profile.cs
@@ -284,7 +284,7 @@ namespace FASTER.Models
             get => ProfileCfgArrays.TacticalPingStrings[tacticalPing];
             set
             {
-                tacticalPing = (int)Array.IndexOf(ProfileCfgArrays.TacticalPingStrings, value);
+                tacticalPing = Array.IndexOf(ProfileCfgArrays.TacticalPingStrings, value);
                 RaisePropertyChanged("TacticalPing");
             }
         }
